made a lot of progress for player movement
