+++ backup_logic.txt
@@ -1,0 +1,58 @@
+if keys.pressed(KeyCode::W) {
+            for (mut player_anim, mut movement, mut pos ) in &mut aseprites.p0().iter_mut() {
+
+                *player_anim = AsepriteAnimation::from(sprites::Player::tags::PLAYER_UP);
+                // *movement = PlayerMovement::Up;
+                // direction.y += 1.;
+
+            }
+            direction.y += 1.;
+
+        }else if keys.just_released(KeyCode::W) {
+            for (mut player_anim, mut movement, mut pos ) in &mut aseprites.p0().iter_mut() {
+                *player_anim = AsepriteAnimation::from(sprites::Player::tags::PLAYER_UP_IDLE);
+            }
+        }
+
+        if keys.just_pressed(KeyCode::A) {
+            for (mut player_anim, mut movement, mut pos ) in &mut aseprites.p0().iter_mut() {
+                *player_anim = AsepriteAnimation::from(sprites::Player::tags::PLAYER_LEFT);
+                //*movement = PlayerMovement::Left;
+            }
+            direction.x -= 1.;
+        }else if keys.just_released(KeyCode::A) {
+            for (mut player_anim, mut movement, mut pos ) in &mut aseprites.p0().iter_mut() {
+                *player_anim = AsepriteAnimation::from(sprites::Player::tags::PLAYER_LEFT_IDLE);
+            }
+        }
+
+        if keys.just_pressed(KeyCode::D) {
+            for (mut player_anim, mut movement, mut pos ) in &mut aseprites.p0().iter_mut() {
+                *player_anim = AsepriteAnimation::from(sprites::Player::tags::PLAYER_RIGHT);
+                //*movement = PlayerMovement::Right;
+            }
+            direction.x += 1.;
+        }else if keys.just_released(KeyCode::D) {
+            for (mut player_anim, mut movement, mut pos ) in &mut aseprites.p0().iter_mut() {
+                *player_anim = AsepriteAnimation::from(sprites::Player::tags::PLAYER_RIGHT_IDLE);
+            }
+        }
+
+        if keys.just_pressed(KeyCode::S) {
+            for (mut player_anim, mut movement, mut pos ) in &mut aseprites.p0().iter_mut() {
+                *player_anim = AsepriteAnimation::from(sprites::Player::tags::PLAYER_DOWN);
+                //*movement = PlayerMovement::Down;
+            }
+            direction.y -= 1.;
+        }else if keys.just_released(KeyCode::S) {
+            for (mut player_anim, mut movement, mut pos ) in &mut aseprites.p0().iter_mut() {
+                *player_anim = AsepriteAnimation::from(sprites::Player::tags::PLAYER_DOWN_IDLE);
+            }
+        }
+
+        let move_speed = 45.;
+        let move_delta = direction * move_speed * time.delta_seconds();
+
+        for (mut player_anim, mut movement, mut pos ) in &mut aseprites.p0().iter_mut() {
+            pos.translation += move_delta.extend(0.);
+        }
